reorganized files and clean-up
