+++ .gitignore
@@ -9,6 +9,10 @@ openoffload/cpp/framework/openoffload.grpc.pb.cc
 openoffload/cpp/framework/openoffload.grpc.pb.h
 openoffload/cpp/framework/openoffload.pb.cc
 openoffload/cpp/framework/openoffload.pb.h
+openoffload/cpp/framework/src/openoffload.grpc.pb.cc
+openoffload/cpp/framework/src/openoffload.grpc.pb.h
+openoffload/cpp/framework/src/openoffload.pb.cc
+openoffload/cpp/framework/src/openoffload.pb.h
 openoffload/cpp/*.cc
 openoffload/cpp/*.h
 openoffload/cpp/tests/openoffload.grpc.pb.cc
